sfdfg

+++ crm/satchi-crm-core/src/main/java/com/fongwell/satchi/crm/core/category/service/CategoryServiceImpl.java
@@ -1,11 +1,14 @@
 package com.fongwell.satchi.crm.core.category.service;
 
+import com.fongwell.base.rest.Payload;
 import com.fongwell.satchi.crm.core.category.domain.aggregate.Category;
 import com.fongwell.satchi.crm.core.category.dto.CategoryData;
 import com.fongwell.satchi.crm.core.category.repository.CategoryRepository;
 import com.fongwell.satchi.crm.core.common.AbstractWriteService;
 import com.fongwell.satchi.crm.core.common.error.DataNotFoundException;
 import com.fongwell.satchi.crm.core.common.error.DuplicateParameterException;
+import com.fongwell.satchi.crm.core.product.query.ProductQueryMapper;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
@@ -25,6 +28,9 @@ public class CategoryServiceImpl extends AbstractWriteService<Category, Long, Ca
 
     @Resource(name = "topCategoryService")
     private TopCategoryService topCategoryService;
+    
+    @Autowired
+    ProductQueryMapper productQueryMapper;
 
     public CategoryServiceImpl() {
         super(Category.class);
@@ -50,14 +56,30 @@ public class CategoryServiceImpl extends AbstractWriteService<Category, Long, Ca
     }
 
     @Override
-    public void onDelete(Collection<Long> ids) {
-
-        List<Category> all = categoryRepository.findAll(ids);
-        for (Category next : all) {
-            next.onDelete();
-            categoryRepository.save(next);
-        }
-        topCategoryService.deleteByParent(ids);
+    public Payload onDelete(Collection<Long> ids) {
+    	System.out.println("CategoryServiceImpl.onDelete:");
+    	int count = 0;
+    	for (Long id : ids) {
+    		System.out.println("id:"+id);
+    		int counts = productQueryMapper.findProduct(id);
+    		System.out.println("有多少条商品数据："+counts);
+    		count+=counts;
+    	}
+    	System.out.println("最后这个分类有里有没有商品:"+count);
+    	
+    	if(count==0) {
+    		 List<Category> all = categoryRepository.findAll(ids);
+    	        for (Category next : all) {
+    	            next.onDelete();
+    	            categoryRepository.save(next);
+    	        }
+    	        topCategoryService.deleteByParent(ids);
+        		System.out.println("---------------------------------200");
+    	        return new Payload("200");
+    	}else {
+    		System.out.println("---------------------------------206");
+    		return new Payload("206");
+    	}
 
     }
